Fixe -- Переименован пакет без использования больших букв в названии. -- Разделены классы на разные файлы. -- Итератору присвоен интерфейс Serializable, из класса Logic этот же интерфейс убран. -- Убраны все public'и у констант, но static у констант Logic оставлены. --
