+++ src/main/java/ru/te3ka_programm/Main.java
@@ -1,0 +1,17 @@
+//***********************//
+//****Program author*****//
+//******Te3Ka_PaynE******//
+//*79811131773@yandex.ru*//
+//***********************//
+
+package ru.te3ka_programm;
+
+import java.io.*;
+
+public class Main {
+    public static void main(String[] args) {
+        Logic logic = new Logic();
+
+        logic.mainMenu();
+    }
+}
